fix: move permissions fragment impl to prompts
